@@ -24,50 +24,46 @@
     ]
   },
   "dependencies": {
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/styles": "^4.11.4",
-    "framer-motion": "^5.0.1",
-    "gsap": "^3.8.0",
-    "isomorphic-unfetch": "^3.1.0",
+    "@emotion/cache": "^11.10.3",
+    "@emotion/react": "^11.10.4",
+    "@emotion/server": "^11.10.0",
+    "@emotion/styled": "^11.10.4",
+    "@mui/icons-material": "^5.10.3",
+    "@mui/material": "^5.10.3",
+    "framer-motion": "^7.2.1",
     "lax.js": "^2.0.3",
-    "next": "^12.1.5",
-    "next-pwa": "^5.4.0",
-    "next-seo": "^4.28.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "next": "^12.2.5",
+    "next-pwa": "^5.6.0",
+    "next-seo": "^5.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-error-boundary": "^3.1.4",
-    "react-intersection-observer": "^8.32.2",
+    "react-intersection-observer": "^9.4.0",
     "react-overdrive": "^0.0.10",
-    "react-plx": "^1.3.17",
+    "react-plx": "^2.0.1",
     "react-progressive-image": "^0.6.0",
-    "react-query": "^3.31.0",
-    "react-spring": "^9.3.0",
-    "react-transition-group": "^4.4.2",
-    "styled-components": "^5.3.3",
+    "react-query": "^3.39.2",
+    "styled-components": "^5.3.5",
     "styled-media-query": "^2.1.2",
     "typed.js": "^2.0.12",
-    "use-axios-client": "^2.0.0",
-    "vconsole": "^3.9.4"
+    "vconsole": "^3.14.6"
   },
   "devDependencies": {
     "@types/material-ui": "^0.21.12",
     "@types/next": "^9.0.0",
-    "@types/node": "^16.11.6",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
-    "@types/react-transition-group": "^4.4.4",
-    "@types/styled-components": "^5.1.15",
-    "@types/styled-jsx": "^3.4.4",
-    "@zeit/next-bundle-analyzer": "^0.1.2",
-    "babel-plugin-styled-components": "^1.13.3",
-    "husky": "^7.0.4",
-    "lint-staged": "^11.2.6",
-    "next-compose-plugins": "^2.2.1",
-    "prettier": "^2.4.1",
-    "typescript": "^4.4.4"
+    "@types/node": "^18.7.14",
+    "@types/react": "^18.0.17",
+    "@types/react-dom": "^18.0.6",
+    "@types/styled-components": "^5.1.26",
+    "babel-plugin-styled-components": "^2.0.7",
+    "eslint": "^8.23.0",
+    "eslint-config-next": "^12.2.5",
+    "husky": "^8.0.1",
+    "lint-staged": "^13.0.3",
+    "prettier": "^2.7.1",
+    "typescript": "^4.8.2"
   },
   "engines": {
-    "node": "^14.x"
+    "node": "^16.x"
   }
 }
