@@ -1,17 +1,26 @@
-import React from 'react'
-import App from 'next/app'
+import { useState, useEffect } from 'react'
 import Head from 'next/head'
+import { AppProps } from 'next/app'
 import Error from 'next/error'
 import { ErrorBoundary } from 'react-error-boundary'
 import { createGlobalStyle } from 'styled-components'
 import AtomSpinner from '../components/common/AtomSpinner'
-import { ThemeProvider } from '@material-ui/styles'
-import CssBaseline from '@material-ui/core/CssBaseline'
-import theme from '../theme'
 import { AnimatePresence, motion } from 'framer-motion'
 import Script from 'next/script'
+import { ThemeProvider } from '@mui/material/styles'
+import CssBaseline from '@mui/material/CssBaseline'
+import { CacheProvider, EmotionCache } from '@emotion/react'
+import theme from '../theme'
+import createEmotionCache from '../lib/createEmotionCache'
 
-const GlobalStyle = createGlobalStyle`
+// Client-side cache, shared for the whole session of the user in the browser.
+const clientSideEmotionCache = createEmotionCache()
+
+interface MyAppProps extends AppProps {
+  emotionCache?: EmotionCache
+}
+
+const GlobalStyle: any = createGlobalStyle`
   @font-face {
     font-family: 'Operator Mono Light';
     src: url('../static/fonts/OperatorMono-Light.otf');
@@ -30,98 +39,86 @@ const GlobalStyle = createGlobalStyle`
   }
 `
 
-export default class MyApp extends App {
-  state = {
-    loading: true,
-  }
+export default function MyApp(props: MyAppProps) {
+  const {
+    Component,
+    emotionCache = clientSideEmotionCache,
+    pageProps,
+    router,
+  } = props
+  const [loading, setLoading] = useState(true)
 
-  componentDidMount() {
-    // Remove the server-side injected CSS.
-    const jssStyles = document.querySelector('#jss-server-side')
-    if (jssStyles && jssStyles.parentNode) {
-      jssStyles.parentNode.removeChild(jssStyles)
-    }
-
+  useEffect(() => {
     requestAnimationFrame(() => {
-      this.setState({ loading: false })
+      setLoading(false)
     })
-  }
+  }, [])
 
-  ErrorFallback() {
-    return <Error statusCode={0} />
-  }
-
-  render() {
-    const { props } = this as any
-    const { Component, pageProps, router } = props
-    const { loading } = this.state
-
-    return (
-      <>
-        <Head>
-          {/* Use minimum-scale=1 to enable GPU rasterization */}
-          <meta
-            name="viewport"
-            content="minimum-scale=1, initial-scale=1, width=device-width, shrink-to-fit=no, viewport-fit=cover"
-          />
-        </Head>
-
-        {/* Global Site Tag (gtag.js) - Google Analytics */}
-        <Script
-          strategy="afterInteractive"
-          src={`https://www.googletagmanager.com/gtag/js?id=${'G-CK8J4LSY7H'}`}
+  return (
+    <CacheProvider value={emotionCache}>
+      <Head>
+        {/* Use minimum-scale=1 to enable GPU rasterization */}
+        <meta
+          name="viewport"
+          content="minimum-scale=1, initial-scale=1, width=device-width, shrink-to-fit=no, viewport-fit=cover"
         />
-        <Script
-          id="gtag-init"
-          strategy="afterInteractive"
-          dangerouslySetInnerHTML={{
-            __html: `
+      </Head>
+
+      {/* Global Site Tag (gtag.js) - Google Analytics */}
+      <Script
+        strategy="afterInteractive"
+        src={`https://www.googletagmanager.com/gtag/js?id=${'G-CK8J4LSY7H'}`}
+      />
+      <Script
+        id="gtag-init"
+        strategy="afterInteractive"
+        dangerouslySetInnerHTML={{
+          __html: `
             window.dataLayer = window.dataLayer || [];
             function gtag(){dataLayer.push(arguments);}
             gtag('js', new Date());
-            gtag('config', '${'G-CK8J4LSY7H'}', {
+            gtag('config', '${'G-JQFYPE6KT9'}', {
               page_path: window.location.pathname,
             });
           `,
-          }}
-        />
+        }}
+      />
 
-        {/* Google Tag Manager - Global base code */}
-        <Script
-          strategy="afterInteractive"
-          dangerouslySetInnerHTML={{
-            __html: `
+      {/* Google Tag Manager - Global base code */}
+      <Script
+        strategy="afterInteractive"
+        dangerouslySetInnerHTML={{
+          __html: `
             (function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
             new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
             j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
             'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
             })(window,document,'script','dataLayer', '${'GTM-K4HJLGF'}');
           `,
-          }}
-        />
+        }}
+      />
 
-        <ErrorBoundary FallbackComponent={this.ErrorFallback}>
-          <GlobalStyle />
-          <AtomSpinner loading={loading} />
-          {loading && <Component {...pageProps} />}
-          <AnimatePresence>
-            <motion.div
-              initial={{ opacity: 0 }}
-              animate={{ opacity: 1 }}
-              exit={{ opacity: 0, position: 'absolute', width: '100%' }}
-              key={router.route}
-            >
-              <ThemeProvider theme={theme}>
-                {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
-                <CssBaseline />
-                <Component {...pageProps} />
-              </ThemeProvider>
-            </motion.div>
-          </AnimatePresence>
-        </ErrorBoundary>
-      </>
-    )
-  }
+      <ErrorBoundary FallbackComponent={() => <Error statusCode={0} />}>
+        <GlobalStyle />
+        <AtomSpinner loading={loading} />
+        {loading && <Component {...pageProps} />}
+        <AnimatePresence>
+          <motion.div
+            initial={{ opacity: 0 }}
+            animate={{ opacity: 1 }}
+            exit={{ opacity: 0, position: 'absolute', width: '100%' }}
+            key={router.route}
+          >
+            <ThemeProvider theme={theme}>
+              {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
+              <CssBaseline />
+              <Component {...pageProps} />
+            </ThemeProvider>
+          </motion.div>
+        </AnimatePresence>
+      </ErrorBoundary>
+    </CacheProvider>
+  )
 }
 
 //TODO: _app.tsx make performance down in https://web.dev
